requirements.txt
construct
# git+git://github.com/koreno/termenu.git#egg=termenu
python-crontab
requests
